% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogenizationfunctions.R
\name{Buishand_range_test}
\alias{Buishand_range_test}
\title{Buishand range test}
\usage{
Buishand_range_test(data, series = "", na.rm = FALSE)
}
\arguments{
\item{na.rm}{}
}
\description{
Buishand range test
}
