% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogenizationfunctions.R
\name{Craddock_test}
\alias{Craddock_test}
\title{craddock test}
\usage{
Craddock_test(data, col_name1 = "", col_name2 = "", na.rm = FALSE,
  dp = 2, ...)
}
\arguments{
\item{...}{}
}
\description{
craddock test
}
