% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogenizationfunctions.R
\name{pettit_test}
\alias{pettit_test}
\title{Pettit test}
\usage{
pettit_test(data_series, ...)
}
\arguments{
\item{data_series}{A climatological time series.}

\item{...}{Additional arguments to be passed into methods.}
}
\value{
Pettit test results.
}
\description{
Pettit test
}
\author{
Shadrack Kibet
}
