% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogenizationfunctions.R
\name{snht}
\alias{snht}
\title{Standard Normal Homogeneity test(SNHT)}
\usage{
snht(data_series, na.rm = FALSE, dp = 2, ...)
}
\arguments{
\item{data_series}{A climatological time series.}

\item{na.rm}{A logical value indicating whether missing values should be stripped before the computation proceeds.}

\item{dp}{Integer indicating the number of decimal places (round) to be used. By default "dp=2".}

\item{...}{Additional arguments to be passed into methods.}
}
\value{
The SNHT test statistic.
}
\description{
Standard Normal Homogeneity test(SNHT)
}
\author{
Shadrack Kibet
}
