% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogenizationfunctions.R
\name{buishand_range_test}
\alias{buishand_range_test}
\title{Buishand range test}
\usage{
buishand_range_test(data_series, na.rm = FALSE, ...)
}
\arguments{
\item{data_series}{A climatological time series.}

\item{na.rm}{A logical value indicating whether missing values should be stripped before the computation proceeds.}

\item{...}{Additional arguments to be passed into methods.}
}
\value{
The Buishand range test results.
}
\description{
Buishand range test
}
\author{
Shadrack Kibet
}
