% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogenizationfunctions.R
\name{craddock_test}
\alias{craddock_test}
\title{Craddock test}
\usage{
craddock_test(data_col1, data_col2, na.rm = FALSE, dp = 2, ...)
}
\arguments{
\item{data_col1}{A candidate climatological series.}

\item{data_col2}{A homogeneous reference series.}

\item{na.rm}{A logical value indicating whether missing values should be stripped before the computation proceeds.}

\item{dp}{Integer indicating the number of decimal places (round) to be used. By default "dp=2".}

\item{...}{Additional arguments to be passed into methods.}
}
\value{
Craddock test results.
}
\description{
Craddock test
}
\author{
Shadrack Kibet
}
