% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogenizationfunctions.R
\name{snht}
\alias{snht}
\title{Standard Normal Homogeneity test(SNHT)}
\usage{
snht(data, series = "", na.rm = FALSE, dp = 2, ...)
}
\arguments{
\item{...}{}
}
\description{
Standard Normal Homogeneity test(SNHT)
}
