% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogenizationfunctions.R
\name{von_neumann_ratio_test}
\alias{von_neumann_ratio_test}
\title{Von Neumman ratio test}
\usage{
von_neumann_ratio_test(data, series = "", na.rm = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
Von Neumman ratio test
}
